--- one-pager-maker/src/Edit.css
@@ -1,58 +1,0 @@
-.editor-title {
-  display: flex;
-  height: 2vh;
-  width: 100%;
-}
-
-.title-text {
-  width: 100%;
-}
-
-.editor-parameter {
-  display: flex;
-  height: 2vh;
-  width: 100%;
-  display: flex;
-  justify-content: space-between;  
-}
-
-.editor-container {
-  display: flex;
-  height: 80vh;
-  width: 100%;
-}
-
-.markdown-input,
-.markdown-preview {
-  flex: 1;
-  padding: 2rem;
-  overflow-y: auto;
-}
-
-.markdown-input {
-  width: 50%;
-  height: 100%;
-  font-size: 1rem;
-  border: none;
-  outline: none;
-  resize: none;
-  background-color: #f4f4f4;
-  padding: 1rem;
-}
-
-.markdown-preview {
-  width: 50%;
-  height: 100%;
-  background-color: #fff;
-  border-left: 1px solid #ddd;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
-}
-
-.editor-parameter-left {
-  text-align: left;
-  float: left; 
-}
-
-.editor-parameter-right {
-  text-align: right;
-}
