--- one-pager-maker/src/Login.tsx
@@ -1,13 +1,0 @@
-import './List.css'
-
-function List() {
-  return (
-    <>
-    <div>
-      <h1>List</h1>
-    </div>
-   </>
-  )
-}
-
-export default List
